Lots of work unifying ideas. Maybe alpha build.
Adds several new modules and features in almost every API
from the channel module's new ping() and pingf() methods
to the shell modules fancy bounce() and echo(). There is
a logical structure to the command syntax, along with a
real parser and analyzer to go with it. Speed could be better,
but that will come after some profiling and lots of thought
about options. For now, it works, and it's almost cohesive.

Soon!